0.32.89

+++ src/main/java/International_Trade_Union/entity/blockchain/DataShortBlockchainInformation.java
@@ -2,19 +2,21 @@ package International_Trade_Union.entity.blockchain;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class DataShortBlockchainInformation {
     private long size;
     private boolean isValidation;
     private long hashCount;
-    private double staking;
+    private BigDecimal staking;
     private long transactions;
     private int bigRandomNumber;
 
     public DataShortBlockchainInformation() {
     }
 
-    public DataShortBlockchainInformation(long size, boolean isValidation, long hashCount, double staking,  long transactions, int bigRandomNumber) {
+    public DataShortBlockchainInformation(long size, boolean isValidation, long hashCount, BigDecimal staking,  long transactions, int bigRandomNumber) {
         this.size = size;
         this.isValidation = isValidation;
         this.hashCount = hashCount;
