+++ src/main/java/International_Trade_Union/simulation/AddapterAccountSimulationToAccount.java
@@ -4,6 +4,7 @@ package International_Trade_Union.simulation;
 
 import International_Trade_Union.model.Account;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -13,7 +14,7 @@ public class AddapterAccountSimulationToAccount {
 
         for (AccountSimulation accountSimulation :accountSimulations) {
 
-            Account account = new Account(accountSimulation.getPublicKey(), accountSimulation.getDigitalDollarBalance());
+            Account account = new Account(accountSimulation.getPublicKey(), BigDecimal.valueOf( accountSimulation.getDigitalDollarBalance()));
             accounts.add(account);
         }
         return accounts;
